feat: extend GraphQL schema for template management
- Add template query fields (default_attributes, configuration, attributes)
- Add template relationship resolvers (attributes, steps)
- Add guide_template_mutation module to GraphQL exports
- Update schema definitions for new template features

The GuideTemplateObject now exposes the full template editing
capabilities through GraphQL including attribute management.

+++ workspace/core/server/src/datastores/content/guides.rs
@@ -1348,7 +1348,7 @@ impl GuidesDataStore {
         &self,
         metadata_id: &Uuid,
         version: i32,
-        step_id: i32,
+        step_id: i64,
     ) -> Result<(), Error> {
         let mut connection = self.pool.get().await?;
         let txn = connection.transaction().await?;
@@ -1364,7 +1364,7 @@ impl GuidesDataStore {
         &self,
         metadata_id: &Uuid,
         version: i32,
-        step_ids: &[i32],
+        step_ids: &[i64],
     ) -> Result<(), Error> {
         let mut connection = self.pool.get().await?;
         let txn = connection.transaction().await?;
